feat: SimulationConfig type mismatched resolved. #11

+++ ThermoTracker/Configurations/SimulationConfig.cs
@@ -5,5 +5,5 @@ public class SimulationConfig
     public int UpdateIntervalMs { get; set; }
     public int DataHistorySize { get; set; }
     public int MovingAverageWindow { get; set; }
-    public int AnomalyThreshold { get; set; }
+    public decimal AnomalyThreshold { get; set; }
 }
